backend/requirements.txt
Django
django-ninja
python-decouple
django-ninja-jwt[crypto]
dj_database_url
pydantic[email]
django-cors-headers
django-allauth
nanoid
requests
djangorestframework-simplejwt